Organize events components by creating component folder and page for it
